pipeline/requirements.txt
boto3
pandas
python-dotenv
redshift_connector